Add order management, refactor import, and enhance API
- Implement order creation, retrieval, and summary features across application, domain, and infrastructure layers
- Refactor customer import to use IFormFile and update related logic
- Update Customer entity to use DateOnly for SignupDate
- Register new services for DI in Application and Infrastructure
- Extend API with order endpoints, Swagger/OpenAPI, and CORS
- Add EF Core migrations and update project dependencies

+++ Domain/Orders/OrderItem.cs
@@ -2,10 +2,10 @@
 
 public class OrderItem
 {
-	public int Id { get; set; }
-	public int OrderId { get; set; }
-	public string Description { get; set; } = string.Empty;
-	public string Sku { get; set; } = string.Empty;
-	public float UnitPrice { get; set; }
-	public int Quantity { get; set; }
+    public int Id { get; set; }
+    public int OrderId { get; set; }
+    public string Description { get; set; } = string.Empty;
+    public string Sku { get; set; } = string.Empty;
+    public decimal UnitPrice { get; set; }
+    public int Quantity { get; set; }
 }
